Disable autodetection, to fix matlab test breakage.
When highlightjs-octave is included in a highlight.js repo, its tests fail, because a file it expects to be detected as "matlab" is now detected as "octave". I think that's because my octave language definition is taking over the matlab matching. This attempts to fix that by disabling autodetection. (Matlab should probably take precedence over Octave in language autodetection anyway, because it's a much more popular language.)